Practica cf

+++ codeforces/Codeforces Round 590 (Div. 3)/A.cpp
@@ -14,12 +14,11 @@ int main(){
     int t; cin >> t;
     while(t--){
         int n; cin >> n;
-        int64_t sum = 0;
+        int sum = 0;
         for(int i = 0; i < n; ++i){
-            int64_t x; cin >> x;
+            int x; cin >> x;
             sum += x;
         }
         cout << (sum + n - 1) / n << '\n';
     }
 }
-
